app_requirements.txt
# django==5.2
python-barcode==0.15.1
djangorestframework==3.16.0